# Validate speedup