--- a/ngraph/core/include/ngraph/interval.hpp
+++ b/ngraph/core/include/ngraph/interval.hpp
@@ -43,7 +43,7 @@ namespace ngraph
         /// \brief The number of elements in the interval. Zero if max < min.
         size_type size() const;
         /// \brief Returns true if the interval has no elements
-        bool empty() const;
+        bool empty() const { return m_min_val == s_max; }
         /// \brief the inclusive lower bound of the interval
         value_type get_min_val() const { return m_min_val; }
         /// \brief Set the inclusive lower bound of the interval
@@ -84,7 +84,7 @@ namespace ngraph
         Interval& operator&=(const Interval& interval);
 
         /// \brief True if this interval includes value
-        bool contains(value_type value) const;
+        bool contains(value_type value) const { return m_min_val <= value && value <= m_max_val; }
         /// \brief True if this interval includes all the values in interval
         bool contains(const Interval& interval) const;
 
@@ -93,7 +93,10 @@ namespace ngraph
 
     protected:
         void canonicalize();
-        static value_type clip(value_type value);
+        static value_type clip(value_type value)
+        {
+            return std::max(value_type(0), std::min(s_max, value));
+        }
         static value_type clip_times(value_type a, value_type b);
         static value_type clip_add(value_type a, value_type b);
         static value_type clip_minus(value_type a, value_type b);