--- a/ngraph/core/src/interval.cpp
+++ b/ngraph/core/src/interval.cpp
@@ -6,6 +6,46 @@
 
 using namespace ngraph;
 
+namespace
+{
+    Interval::value_type clip(Interval::value_type value)
+    {
+        return std::max(Interval::value_type(0), std::min(Interval::s_max, value));
+    }
+
+    Interval::value_type clip_times(Interval::value_type a, Interval::value_type b)
+    {
+        if (a == 0 || b == 0)
+        {
+            return 0;
+        }
+        else if (a == Interval::s_max || b == Interval::s_max)
+        {
+            return Interval::s_max;
+        }
+        else
+        {
+            return a * b;
+        }
+    }
+    Interval::value_type clip_add(Interval::value_type a, Interval::value_type b)
+    {
+        return (a == Interval::s_max || b == Interval::s_max) ? Interval::s_max : a + b;
+    }
+    Interval::value_type clip_minus(Interval::value_type a, Interval::value_type b)
+    {
+        if (a <= b)
+        {
+            return 0;
+        }
+        if (a == Interval::s_max)
+        {
+            return Interval::s_max;
+        }
+        return a - b;
+    }
+} // namespace
+
 void Interval::canonicalize()
 {
     if (m_max_val < m_min_val)
@@ -113,40 +153,6 @@ Interval& Interval::operator&=(const Interval& interval)
 bool Interval::contains(const Interval& interval) const
 {
     return contains(interval.m_min_val) && contains(interval.m_max_val);
-}
-
-Interval::value_type Interval::clip_add(value_type a, value_type b)
-{
-    return (a == s_max || b == s_max) ? s_max : a + b;
-}
-
-Interval::value_type Interval::clip_minus(value_type a, value_type b)
-{
-    if (a <= b)
-    {
-        return 0;
-    }
-    if (a == s_max)
-    {
-        return s_max;
-    }
-    return a - b;
-}
-
-Interval::value_type Interval::clip_times(value_type a, value_type b)
-{
-    if (a == 0 || b == 0)
-    {
-        return 0;
-    }
-    else if (a == s_max || b == s_max)
-    {
-        return s_max;
-    }
-    else
-    {
-        return a * b;
-    }
 }
 
 constexpr Interval::value_type Interval::s_max;