### Details:
 - *move some method to headers which allow it to be inline*
 - *`PartialShape` c-tors will consume (move) vector of `Dimension`*
 - *initialize map returning by `get_type_info_map` before `main` - purge static function member "mutex"*
 - *find in map and `terminate` if not found `TypeInfo`*

### Tickets:
 - *61261*

### Output:
 - *FakeQuantize::validate_and_infer_types* approximately speedup: 20%