@pelszkow 2021-07-29
> Do you have any measurements which show improvements?
master branch:
```
./unit-test --gtest_filter="IN*smoke_validation*"                                                                                                                                                                                                                           master
Note: Google Test filter = IN*smoke_validation*
[==========] Running 2 tests from 1 test suite.
[----------] Global test environment set-up.
[----------] 2 tests from INTERPRETER
[ RUN      ] INTERPRETER.smoke_validation_reshape
[[[[[[ 1.73853e+06us ]]]]]]
[       OK ] INTERPRETER.smoke_validation_reshape (1739 ms)
[ RUN      ] INTERPRETER.smoke_validation_fake_quantize
[[[[[[ 653971us ]]]]]]
[       OK ] INTERPRETER.smoke_validation_fake_quantize (654 ms)
[----------] 2 tests from INTERPRETER (2393 ms total)

[----------] Global test environment tear-down
[==========] 2 tests from 1 test suite ran. (2393 ms total)
[  PASSED  ] 2 tests.
```

after this PR:
```
./unit-test --gtest_filter="IN*smoke_validation*"                                                                                                                                                                                                                 validate_speedup
Note: Google Test filter = IN*smoke_validation*
[==========] Running 2 tests from 1 test suite.
[----------] Global test environment set-up.
[----------] 2 tests from INTERPRETER
[ RUN      ] INTERPRETER.smoke_validation_reshape
[[[[[[ 1.40525e+06us ]]]]]]
[       OK ] INTERPRETER.smoke_validation_reshape (1405 ms)
[ RUN      ] INTERPRETER.smoke_validation_fake_quantize
[[[[[[ 403533us ]]]]]]
[       OK ] INTERPRETER.smoke_validation_fake_quantize (404 ms)
[----------] 2 tests from INTERPRETER (1809 ms total)

[----------] Global test environment tear-down
[==========] 2 tests from 1 test suite ran. (1809 ms total)
[  PASSED  ] 2 tests.
```
Where tests looks like this:
```
diff --git a/ngraph/test/backend/fake_quantize.in.cpp b/ngraph/test/backend/fake_quantize.in.cpp
index 363e89b31..d5bb49490 100644
--- a/ngraph/test/backend/fake_quantize.in.cpp
+++ b/ngraph/test/backend/fake_quantize.in.cpp
@@ -31,6 +31,31 @@ using namespace ngraph;
 static std::string s_manifest = "${MANIFEST}";
 using TestEngine = test::ENGINE_CLASS_NAME(${BACKEND_NAME});

+NGRAPH_TEST(${BACKEND_NAME}, smoke_validation_fake_quantize)
+{
+    const Shape data_shape{1, 2, 3, 4};
+    const size_t levels = 4;
+    const auto data = std::make_shared<op::Parameter>(element::f32, data_shape);
+    const auto input_low = op::Constant::create(element::f32, Shape{}, {0.f});
+    const auto input_high = op::Constant::create(element::f32, Shape{}, {23.f});
+    const auto output_low = op::Constant::create(element::f32, Shape{}, {2.f});
+    const auto output_high = op::Constant::create(element::f32, Shape{}, {16.f});
+
+    const auto quantize = std::make_shared<op::FakeQuantize>(
+        data, input_low, input_high, output_low, output_high, levels);
+
+    const auto start_time = std::chrono::high_resolution_clock::now();
+    constexpr auto validation_rounds = 1000000ul;
+    for (size_t i = 0; i < validation_rounds; ++i)
+    {
+        quantize->validate_and_infer_types();
+    }
+    const float duration = std::chrono::duration_cast<std::chrono::microseconds>(
+                               std::chrono::high_resolution_clock::now() - start_time)
+                               .count();
+    std::cerr << "[[[[[[ " << duration << "us ]]]]]]" << std::endl;
+}
+
 namespace
 {
     template <typename T>
diff --git a/ngraph/test/backend/reshape.in.cpp b/ngraph/test/backend/reshape.in.cpp
index f49870bf1..cb30f54a7 100644
--- a/ngraph/test/backend/reshape.in.cpp
+++ b/ngraph/test/backend/reshape.in.cpp
@@ -32,6 +32,26 @@ static string s_manifest = "${MANIFEST}";

 using TestEngine = test::ENGINE_CLASS_NAME(${BACKEND_NAME});

+NGRAPH_TEST(${BACKEND_NAME}, smoke_validation_reshape)
+{
+    Shape shape_a{2, 2, 3};
+    auto A = make_shared<op::Parameter>(element::f32, shape_a);
+    Shape shape_r{12};
+    auto r = make_shared<op::v1::Reshape>(
+        A, op::Constant::create(element::u64, {shape_r.size()}, shape_r), false);
+
+    const auto start_time = std::chrono::high_resolution_clock::now();
+    constexpr auto validation_rounds = 1000000ul;
+    for (size_t i = 0; i < validation_rounds; ++i)
+    {
+        r->validate_and_infer_types();
+    }
+    const float duration = std::chrono::duration_cast<std::chrono::microseconds>(
+                               std::chrono::high_resolution_clock::now() - start_time)
+                               .count();
+    std::cerr << "[[[[[[ " << duration << "us ]]]]]]" << std::endl;
+}
+
 NGRAPH_TEST(${BACKEND_NAME}, reshape_t2v)
 {
     Shape shape_a{2, 2, 3};
```